+++ src/components/ReportUpload/ReportUpload.js
@@ -31,6 +31,7 @@ class ReportUpload extends React.Component {
       showCounterUpload: false,
       showNonCounterUpload: false,
     });
+    this.props.onReloadStatistics();
   };
 
   handleFail = (msg) => {
@@ -93,6 +94,7 @@ class ReportUpload extends React.Component {
 ReportUpload.propTypes = {
   handlers: PropTypes.shape(),
   intl: PropTypes.object,
+  onReloadStatistics: PropTypes.func.isRequired,
   stripes: PropTypes.shape().isRequired,
   udpId: PropTypes.string,
 };