+++ src/components/Statistics/Counter/ReportInfo/ReportInfo.js
@@ -1,14 +1,10 @@
-import _ from 'lodash';
+import { get } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl, FormattedMessage } from 'react-intl';
-import {
-  Button,
-  Icon,
-  KeyValue,
-  MenuSection,
-} from '@folio/stripes/components';
+import { Button, Icon, KeyValue, MenuSection } from '@folio/stripes/components';
 import reportDownloadTypes from '../../../../util/data/reportDownloadTypes';
+import isSushiWarningCode from '../../../../util/isSushiWarningCode';
 
 class ReportInfo extends React.Component {
   onClickDownloadRawReport = (release) => {
@@ -124,63 +120,107 @@ class ReportInfo extends React.Component {
   };
 
   manualEditedText() {
+    const editReason = ` ${get(this.props.report, 'editReason', '-')}`;
     return (
       <>
         <FormattedMessage id="ui-erm-usage.general.manualChanges.infoText" />
         <br />
         <FormattedMessage id="ui-erm-usage.general.editReason" />
-        {_.get(this.props.report, 'editReason', '-')}
+        {editReason}
       </>
     );
   }
 
+  translateErrorCodes = (val) => {
+    const { intl } = this.props;
+    let label;
+    if (isSushiWarningCode(val)) {
+      label = `${intl.formatMessage({
+        id: 'ui-erm-usage.report.error.1',
+      })} (${val})`;
+    } else {
+      const id = `ui-erm-usage.report.error.${val}`;
+      label = `${intl.formatMessage({
+        id,
+      })} (${val})`;
+    }
+    return `${intl.formatMessage({
+      id: 'ui-erm-usage.report.error.sushiException',
+    })}: ${label}`;
+  };
+
+  adaptSushiFailedInfo(failedReason) {
+    if (
+      failedReason.includes('Number') &&
+      failedReason.includes('Severity') &&
+      failedReason.includes('Message')
+    ) {
+      const errorCode = failedReason.match('Number=(.*), Severity=')[1];
+      return this.translateErrorCodes(errorCode);
+    } else {
+      return failedReason;
+    }
+  }
+
   render() {
-    const { report, retryThreshold } = this.props;
+    const { intl, report, retryThreshold } = this.props;
 
     const failInfo = !report.failedReason ? null : (
       <KeyValue
-        label={this.props.intl.formatMessage({
+        data-test-report-failed-reason
+        label={intl.formatMessage({
           id: 'ui-erm-usage.general.info',
         })}
-        value={report.failedReason}
+        value={this.adaptSushiFailedInfo(report.failedReason)}
       />
     );
 
     const failedAttempts = !report.failedAttempts ? null : (
       <KeyValue
-        label={this.props.intl.formatMessage({
+        label={intl.formatMessage({
           id: 'ui-erm-usage.report.action.failedAttempts',
         })}
-        value={`${report.failedAttempts} (Max attempts: ${retryThreshold})`}
+        value={intl.formatMessage(
+          {
+            id: 'ui-erm-usage.statistics.report.info.maxAttempts',
+          },
+          {
+            current: report.failedAttempts,
+            max: retryThreshold,
+          }
+        )}
       />
     );
 
-    const displayManualEdited = (
-      report.reportEditedManually ?
-        <KeyValue
-          data-test-custom-reports-edited-manually
-          label={this.props.intl.formatMessage({ id: 'ui-erm-usage.general.manualChanges' })}
-          value={this.manualEditedText()}
-        />
-        : '');
+    const displayManualEdited = report.reportEditedManually ? (
+      <KeyValue
+        data-test-custom-reports-edited-manually
+        label={intl.formatMessage({
+          id: 'ui-erm-usage.general.manualChanges',
+        })}
+        value={this.manualEditedText()}
+      />
+    ) : (
+      ''
+    );
 
     const headerSection = (
       <MenuSection
         id="menu-actions"
-        label={this.props.intl.formatMessage({
+        label={intl.formatMessage({
           id: 'ui-erm-usage.general.report',
         })}
         labelTag="h3"
       >
         <KeyValue label="Usage data provider" value={this.props.udpLabel} />
         <KeyValue
-          label={this.props.intl.formatMessage({
+          label={intl.formatMessage({
             id: 'ui-erm-usage.general.type',
           })}
           value={report.reportName}
         />
         <KeyValue
-          label={this.props.intl.formatMessage({
+          label={intl.formatMessage({
             id: 'ui-erm-usage.general.date',
           })}
           value={report.yearMonth}
@@ -199,7 +239,7 @@ class ReportInfo extends React.Component {
     const actionSection = (
       <MenuSection
         id="menu-actions"
-        label={this.props.intl.formatMessage({
+        label={intl.formatMessage({
           id: 'ui-erm-usage.general.actions',
         })}
         labelTag="h3"