+++ src/settings/MaxFailedAttempts/MaxFailedAttempts.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { Col, Pane, Row, TextField } from '@folio/stripes/components';
+import { Col, Row, TextField } from '@folio/stripes/components';
 import { Field } from 'redux-form';
 import { ConfigManager } from '@folio/stripes/smart-components';
 import { MAX_FAILED_ATTEMPTS } from '../../util/constants';
 
 class MaxFailedAttempts extends React.Component {
   static propTypes = {
-    stripes: PropTypes.object
+    stripes: PropTypes.object,
   };
 
   constructor(props) {
@@ -16,28 +16,13 @@ class MaxFailedAttempts extends React.Component {
     this.configManager = props.stripes.connect(ConfigManager);
   }
 
-  renderNotPresent = () => {
-    return (
-      <Pane
-        defaultWidth="fill"
-        fluidContentWidth
-        paneTitle="Required interface not present"
-      >
-        <div>
-          The interface usage-harvester is needed to start the harvester, but it
-          is not present.
-        </div>
-      </Pane>
-    );
-  };
-
-  getInitialValues = settings => {
+  getInitialValues = (settings) => {
     let loadedValues = {};
     let value;
     try {
       value = settings.length === 0 ? MAX_FAILED_ATTEMPTS : settings[0].value;
       loadedValues = {
-        maxFailedAttempts: value
+        maxFailedAttempts: value,
       };
     } catch (e) {} // eslint-disable-line no-empty
     return loadedValues;
@@ -65,7 +50,9 @@ class MaxFailedAttempts extends React.Component {
                   type="number"
                   id="maxFailedAttempts"
                   name="maxFailedAttempts"
-                  label="Number failed attempts"
+                  label={
+                    <FormattedMessage id="ui-erm-usage.settings.section.number.failed" />
+                  }
                 />
               </Col>
             </Row>