+++ src/util/data/harvestingStatusOptions.js
@@ -1,4 +1,4 @@
 export default [
-  { value: 'active', label: 'Active' },
-  { value: 'inactive', label: 'Inactive' }
+  { label: 'ui-erm-usage.general.status.active', value: 'active' },
+  { label: 'ui-erm-usage.general.status.inactive', value: 'inactive' }
 ];