+++ test/bigtest/interactors/udp-details-page.js
@@ -149,6 +149,7 @@ const counterevent = { dataTransfer: { files: [counterfile], types: ['Files'] } 
   static defaultScope = '[class="report-info-failed"]';
   downloadJsonXmlButton = scoped('button[id="download-json-xml-button"]');
   deleteButton = scoped('button[id="delete-report-button"]');
+  failedReason = scoped('[data-test-report-failed-reason]');
 }
 
 @interactor class CustomReports {
@@ -156,6 +157,10 @@ const counterevent = { dataTransfer: { files: [counterfile], types: ['Files'] } 
 
   instances = collection('[role=row]');
   clickFirstRow = clickable('[id*=custom-report-button-]');
+}
+
+@interactor class CustomReportAccordion2000 {
+  static defaultScope = '#accordion-toggle-button-2000';
 }
 
 @interactor class CustomReportInfo {
@@ -220,6 +225,18 @@ const counterevent = { dataTransfer: { files: [counterfile], types: ['Files'] } 
   tagSelection = scoped('#input-tag-input');
 }
 
+@interactor class ActionsDropdown {
+  click = clickable('[data-test-pane-header-actions-button]');
+}
+
+@interactor class RefreshStatsButton {
+  static defaultScope = ('#clickable-refresh-statistics');
+}
+
+@interactor class EditUDPButton {
+  static defaultScope = ('#clickable-edit-udp');
+}
+
 export default @interactor class UDPDetailsPage {
   static defaultScope = '#pane-udpdetails';
   title = text('[data-test-header-title]');
@@ -270,4 +287,9 @@ export default @interactor class UDPDetailsPage {
   urlInputError = new UrlInputError();
 
   saveNonCounterButton = new SaveNonCounterButton();
+  customReportAccordion2000 = new CustomReportAccordion2000();
+
+  actionsButton = new ActionsDropdown();
+  refreshStatsButton = new RefreshStatsButton();
+  editUDPButton = new EditUDPButton();
 }